Add files via upload
This PDF is the marked up schematic to align with AS-BUILT units.
Please refer to the notes and comments embedded in the PDF.
These will be incorporated in the final published design files. 
There are some parts such as a ultracap/battery that have been identified as difficult to get. As a result the final design may see small footprint changes to correctly adapt to allow a replacement to be placed on the board. There should NOT be any significant layout/trace changes on the board.